requirements.txt
addict==2.4.0
albumentations==1.3.0
backcall==0.2.0
certifi==2022.9.24
chainer==7.8.1
charset-normalizer==2.0.12
cupy @ file:///home/conda/feedstock_root/build_artifacts/cupy_1630031376637/work
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
decord==0.6.0
docopt==0.6.2
fastrlock==0.8.1
filelock==3.4.1
googledrivedownloader==0.4
grad-cam==1.4.6
idna==3.4
imageio==2.15.0
importlib-resources==5.4.0
ipython==7.16.3
ipython-genutils==0.2.0
isodate==0.6.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mmcv-full==1.7.0
# Editable install with no version control (mmdet==2.12.0)
-e /home/data2/lzc/WSIS-Benchmark/code/mmdection_edit
networkx==2.5.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1626681920064/work
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipreqs==0.4.11
prompt-toolkit==3.0.32
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pycocotools==2.0.6
pydensecrf==1.0rc3
Pygments==2.13.0
pynvml==11.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.1.1
PyYAML==4.2b4
qudida==0.0.4
rdflib==5.0.0
requests==2.27.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
sklearn==0.0.post1
skorch==0.11.0
tabulate==0.8.10
tensorboardX==2.5.1
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.10.0+cu111
torch-cluster==1.5.9
torch-geometric==2.0.3
torch-scatter==2.0.9
torch-sparse==0.6.12
torchaudio==0.10.0+rocm4.1
torchcam==0.3.2
torchsummary==1.5.1
torchvision==0.11.0+cu111
tqdm==4.64.1
traitlets==4.3.3
ttach==0.0.3
typing_extensions==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
wfdb==3.4.1
yacs==0.1.8
yapf==0.32.0
yarg==0.1.9
zipp==3.6.0
